+++ src/lib/components/lights/AmbientLight.svelte
@@ -2,11 +2,15 @@
 	import { setup } from '../../utils/context';
 	import * as THREE from 'three';
 
+	/** @type {THREE.AmbientLight} */
+	export let light;
+
 	/** @type {string | number | THREE.Color} */
 	export let color = 0xffffff;
 	export let intensity = 1;
 
 	const { root, self } = setup(new THREE.AmbientLight(color, intensity));
+	light = self;
 
 	$: {
 		self.color.set(color);