+++ src/lib/components/objects/Bone.svelte
@@ -4,11 +4,15 @@
 	import { transform } from '../../utils/object.js';
 	import * as defaults from '../../utils/defaults.js';
 
+	/** @type {THREE.Bone} */
+	export let bone;
+
 	export let position = defaults.position;
 	export let rotation = defaults.rotation;
 	export let scale = defaults.scale;
 
 	const { root, self } = setup(new THREE.Bone());
+	bone = self;
 
 	$: {
 		transform(self, position, rotation, scale);